+++ include/public/LibGraphics/ocr/OcrResult.hpp
@@ -5,6 +5,6 @@ using LibGraphics::Image;
 namespace LibGraphics::Ocr {
     struct OcrResult {
         std::string text;
-        float confidence;
+        double confidence;
     };
 }
